+++ pkg/scheduler/watchers/clusterschedulingpolicysnapshot/suite_test.go
@@ -21,7 +21,7 @@ import (
 
 	fleetv1beta1 "go.goms.io/fleet/apis/placement/v1beta1"
 	"go.goms.io/fleet/pkg/scheduler/queue"
-	"go.goms.io/fleet/pkg/utils"
+	"go.goms.io/fleet/test/utils/keycollector"
 )
 
 var (
@@ -29,7 +29,7 @@ var (
 	hubClient    client.Client
 	ctx          context.Context
 	cancel       context.CancelFunc
-	keyCollector *utils.SchedulerWorkqueueKeyCollector
+	keyCollector *keycollector.SchedulerWorkqueueKeyCollector
 )
 
 func TestAPIs(t *testing.T) {
@@ -79,7 +79,7 @@ var _ = BeforeSuite(func() {
 	Expect(err).ToNot(HaveOccurred(), "Failed to set up controller with controller manager")
 
 	// Start the key collector.
-	keyCollector = utils.NewSchedulerWorkqueueKeyCollector(schedulerWorkQueue)
+	keyCollector = keycollector.NewSchedulerWorkqueueKeyCollector(schedulerWorkQueue)
 	go func() {
 		keyCollector.Run(ctx)
 	}()
