+++ pkg/utils/test_util.go
@@ -6,11 +6,9 @@ Licensed under the MIT license.
 package utils
 
 import (
-	"context"
 	"fmt"
 	"os"
 	"strings"
-	"sync"
 
 	"github.com/onsi/gomega/format"
 	v1 "k8s.io/api/core/v1"
@@ -22,8 +20,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/yaml"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/tools/record"
-
-	"go.goms.io/fleet/pkg/scheduler/queue"
 )
 
 var (
@@ -139,67 +135,4 @@ func (matcher AlreadyExistMatcher) FailureMessage(actual interface{}) (message s
 // NegatedFailureMessage builds an error message.
 func (matcher AlreadyExistMatcher) NegatedFailureMessage(actual interface{}) (message string) {
 	return format.Message(actual, "not to be already exist")
-}
-
-// SchedulerWorkqueueKeyCollector helps collect keys from a scheduler work queue for testing
-// purposes.
-type SchedulerWorkqueueKeyCollector struct {
-	schedulerWorkqueue queue.ClusterResourcePlacementSchedulingQueue
-	// Uses a mutex to guard against concurrent access; for simplicity reasons, the struct
-	// uses a regular map rather than its currency safe variant.
-	lock          sync.Mutex
-	collectedKeys map[string]bool
-}
-
-// NewSchedulerWorkqueueKeyCollector returns a new SchedulerWorkqueueKeyCollector.
-func NewSchedulerWorkqueueKeyCollector(wq queue.ClusterResourcePlacementSchedulingQueue) *SchedulerWorkqueueKeyCollector {
-	return &SchedulerWorkqueueKeyCollector{
-		schedulerWorkqueue: wq,
-		collectedKeys:      make(map[string]bool),
-	}
-}
-
-// Run runs the SchedulerWorkqueueKeyCollector.
-func (kc *SchedulerWorkqueueKeyCollector) Run(ctx context.Context) {
-	go func() {
-		for {
-			key, closed := kc.schedulerWorkqueue.NextClusterResourcePlacementKey()
-			if closed {
-				break
-			}
-
-			kc.lock.Lock()
-			kc.collectedKeys[string(key)] = true
-			kc.schedulerWorkqueue.Done(key)
-			kc.schedulerWorkqueue.Forget(key)
-			kc.lock.Unlock()
-		}
-	}()
-
-	<-ctx.Done()
-}
-
-// IsPresent returns whether a given key is has been collected.
-func (kc *SchedulerWorkqueueKeyCollector) IsPresent(key string) bool {
-	kc.lock.Lock()
-	defer kc.lock.Unlock()
-
-	_, ok := kc.collectedKeys[key]
-	return ok
-}
-
-// Reset clears all the collected keys.
-func (kc *SchedulerWorkqueueKeyCollector) Reset() {
-	kc.lock.Lock()
-	defer kc.lock.Unlock()
-
-	kc.collectedKeys = make(map[string]bool)
-}
-
-// Len returns the count of collected keys.
-func (kc *SchedulerWorkqueueKeyCollector) Len() int {
-	kc.lock.Lock()
-	defer kc.lock.Unlock()
-
-	return len(kc.collectedKeys)
 }
